agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "damai-agents"
version = "0.1.0"
description = "Agents chat service for damai-go"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "grpcio>=1.71.0",
  "grpcio-tools>=1.71.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "langchain>=1.0.0",
  "langchain-mcp-adapters>=0.1.9",
  "langchain-openai>=1.0.1",
  "langgraph>=1.0.2",
  "mcp[cli]>=1.20.0",
  "pydantic-settings>=2.11.0",
  "pymysql>=1.1.1",
  "pytest>=8.3.0",
  "python-dotenv>=1.1.1",
  "redis>=5.2.0",
  "uvicorn>=0.34.0",
]

[project.scripts]
damai-mcp-server = "app.mcp_server.server:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
